Promise.race()
The race method sends requests to the same endpoint hosted at multiple
servers to get back the first/fastest result and uses that.
If the first/fastest result is an error, it hits the catch block, and
you don't get your data (even if the slower ones would be fulfilled
since it stops after the fastest one gets back).